// src/enums/Constants.tsx
export enum ScreenOrientationName {
    portrait = 'portrait',
    landscape = 'landscape'
}

export enum RouteName {
    home = 'Home',
    game = 'Game'
}

export enum ColorName {
    colorPrimary = '#283655',
    colorPrimaryDark = '#1E1F26',
    colorAccent = '#4D648D',
    colorCell = '#D0E1F9',
    colorCellBorder = '#4D648D'
}